Make OBT output errors when --watch is used if an error exists
