+++ lib/origami-build-tools-cli.js
@@ -2,10 +2,10 @@
 
 'use strict';
 
+const path = require('path');
 const argv = require('minimist')(process.argv.slice(2));
 const gulp = require('gulp');
 const tasks = require('./origami-build-tools');
-const watcher = require('./tasks/watch.js');
 const log = require('./helpers/log');
 
 function getLengthOfLongestCommand(commands) {
@@ -54,6 +54,46 @@ function printUsage(commands) {
 	console.log('Full documentation: http://git.io/bBjMNw');
 }
 
+function runWatcher(task, gulp, config) {
+	const watchGlobs = [
+		// include
+		'**/*.js',
+		'**/*.scss',
+		'**/*.mustache',
+		'**/*.json',
+		// exclude
+		'!build/**',
+		'!node_modules/**',
+		'!bower_components/**',
+		'!demos/*',
+		'!demos/local/*',
+		'!origami.json',
+		'!bower.json',
+		'!package.json',
+		'!**/tmp-src.scss'
+	];
+	const watcher = gulp.watch(watchGlobs);
+
+	if (typeof task !== 'function') {
+		return;
+	}
+	watcher.on('ready', function() {
+		log.secondary('Running tasks...');
+		checkTaskResult(task(gulp, config));
+	});
+
+	watcher.on('change', function(event) {
+		log.secondary('File ' + event.path + ' was ' + event.type + ', running tasks...');
+		const fileExtension = path.extname(event.path);
+		if (fileExtension === '.js') {
+			config.watching = 'js';
+		} else if (fileExtension === '.scss') {
+			config.watching = 'sass';
+		}
+		checkTaskResult(task(gulp, config));
+	});
+}
+
 // Mini-hack to make Origami Build Tools support the pseudo standard CLI language of --version
 if (argv.version) {
 	delete argv.version;
@@ -76,29 +116,31 @@ function reportFinished() {
 	log.primary('Finished running ' + argument);
 }
 
+function checkTaskResult(taskResult) {
+	if (typeof taskResult !== 'undefined' && taskResult !== null) {
+		// Check if the function returns a Promise
+		if (taskResult instanceof Promise) {
+			taskResult
+				.then(reportFinished)
+				.catch(reportTaskError);
+		// Check if it's a stream
+		} else if (taskResult.on && taskResult.resume) {
+			taskResult
+				.on('error', reportTaskError)
+				.on('end', reportFinished);
+			// Make sure it reaches 'end'
+			taskResult.resume();
+		}
+	}
+}
+
 if (tasks.isValid(argument)) {
 	const task = tasks[argument];
 
 	if (watch && task.watchable) {
-		watcher.run(task, gulp, argv);
+		runWatcher(task, gulp, argv);
 	} else {
-		const taskResult = task(gulp, argv);
-
-		if (typeof taskResult !== 'undefined' && taskResult !== null) {
-			// Check if the function returns a Promise
-			if (taskResult instanceof Promise) {
-				taskResult
-					.then(reportFinished)
-					.catch(reportTaskError);
-			// Check if it's a stream
-			} else if (taskResult.on && taskResult.resume) {
-				taskResult
-					.on('error', reportTaskError)
-					.on('end', reportFinished);
-				// Make sure it reaches 'end'
-				taskResult.resume();
-			}
-		}
+		checkTaskResult(task(gulp, argv));
 	}
 } else {
 	if (argument) {
